Remove NodeJS code (do not confuse them with "node_modules" content)
I intend use a client-side approach, based in one "json like" file. It will be fastest
than doing requests to GitHub (with a json as the target) and it will be simplest than
using a lot of server-side scripts in a page without server-side. This page is light,
then I think that it will not be a problem to client, unless the client is using a
fridge to access the page (LOL).