Fix wrong scale (in phones) with meta-content-viewport:
I did not know that this meta-tag (`<meta content="viewport" .../>`)
is so important to dimension the page in phones. When I was rebuilding
the page source code, I left to put it later, but I forgot to do this.
It was hard to discover that the problem was its absence, but, at some
time, I throught: "What is if the absence of that tag is making some
difference?"