Fix STUPID bug:
The "expand button", from home/cards, did not update
its icon after the first use.

I really do not know how I did not see it! It was
obvious; it was explicity. I only can see it after
the deploy, while I was browsing the website with
my phone. LOL!