Fix type validation bug (+fix `undefined` when copy an url from cards):
  Startly, I wish of that `copyTextFromInputToClipboard()` could
copy strings, no only content from `input` (or `textarea`)
elements, because of it, I putted the type validation code
after the first trying of copy something (using the
modern Clipboard API, that it support the sopy of strings),
but I given up to do this, because it will require a
fallback `input` element to old browsers.
  Sadly, I forgot to move the type validation code to the
function top after to decide it, what it caused the problem
that I fixed in this commit.